src/pipeline/airflow_dags.py
"""
📅 APACHE AIRFLOW DAGS
Orquestração de pipelines AML
"""
from datetime import datetime, timedelta
from loguru import logger

try:
    from airflow import DAG
    from airflow.operators.python import PythonOperator
    from airflow.operators.bash import BashOperator
    AIRFLOW_AVAILABLE = True
except ImportError:
    AIRFLOW_AVAILABLE = False
    logger.warning("apache-airflow not installed")


# offault args for all os DAGs
default_args = {
    'owner': 'aml-team',
    'depends_on_past': False,
    'email': ['alerts@aml-system.com'],
    'email_on_failure': True,
    'email_on_retry': False,
    'retries': 3,
    'retry_delay': timedelta(minutes=5),
}


def refresh_sanctions_lists():
    """Task: Updates listas of sanções"""
    logger.info("📥 Refreshing sanctions lists...")
    # Import aqui for evitar circular ofpenofncy
    from ..data.sanctions_loader import GlobalSanctionsChecker
    
    checker = GlobalSanctionsChecker()
    checker.refresh_all_lists()
    
    logger.success("✅ Sanctions lists refreshed")


def train_ml_models():
    """Task: Treina/retreina mooflos ML"""
    logger.info("🤖 Training ML models...")
    # Implinentação real treinaria XGBoost/LightGBM/CatBoost
    logger.success("✅ ML models trained")


def generate_daily_report():
    """Task: Generates report diário"""
    logger.info("📊 Generating daily report...")
    logger.success("✅ Daily report generated")


def detect_anomalies_batch():
    """Task: oftects anomalias in batch"""
    logger.info("🔍 Running batch anomaly detection...")
    logger.success("✅ Batch detection complete")


if AIRFLOW_AVAILABLE:
    
    # DAG 1: Daily Sanctions Refresh
    sanctions_refresh_dag = DAG(
        'sanctions_refresh',
        default_args=default_args,
        description='Refresh sanctions lists daily',
        schedule_interval='0 2 * * *',  # 2 AM daily
        start_date=datetime(2024, 1, 1),
        catchup=False,
        tags=['compliance', 'sanctions'],
    )
    
    with sanctions_refresh_dag:
        refresh_task = PythonOperator(
            task_id='refresh_sanctions',
            python_callable=refresh_sanctions_lists,
        )
    
    
    # DAG 2: Weekly ML Moofl Training
    ml_training_dag = DAG(
        'ml_model_training',
        default_args=default_args,
        description='Train/retrain ML models weekly',
        schedule_interval='0 3 * * 0',  # 3 AM every Sunday
        start_date=datetime(2024, 1, 1),
        catchup=False,
        tags=['ml', 'training'],
    )
    
    with ml_training_dag:
        train_task = PythonOperator(
            task_id='train_models',
            python_callable=train_ml_models,
        )
    
    
    # DAG 3: Daily AML Report
    daily_report_dag = DAG(
        'daily_aml_report',
        default_args=default_args,
        description='Generate daily AML report',
        schedule_interval='0 8 * * *',  # 8 AM daily
        start_date=datetime(2024, 1, 1),
        catchup=False,
        tags=['reporting'],
    )
    
    with daily_report_dag:
        report_task = PythonOperator(
            task_id='generate_report',
            python_callable=generate_daily_report,
        )
    
    
    # DAG 4: Horrly Batch oftection
    batch_detection_dag = DAG(
        'batch_anomaly_detection',
        default_args=default_args,
        description='Batch anomaly detection hourly',
        schedule_interval='0 * * * *',  # Every hour
        start_date=datetime(2024, 1, 1),
        catchup=False,
        tags=['detection', 'batch'],
    )
    
    with batch_detection_dag:
        detect_task = PythonOperator(
            task_id='detect_anomalies',
            python_callable=detect_anomalies_batch,
        )
    
    
    # DAG 5: withplex Pipelinand (Multipland tasks)
    complex_pipeline_dag = DAG(
        'complex_aml_pipeline',
        default_args=default_args,
        description='Complete AML pipeline',
        schedule_interval='0 1 * * *',  # 1 AM daily
        start_date=datetime(2024, 1, 1),
        catchup=False,
        tags=['pipeline', 'complete'],
    )
    
    with complex_pipeline_dag:
        # Task 1: Data extraction
        extract = BashOperator(
            task_id='extract_data',
            bash_command='echo "Extracting data from sources..."',
        )
        
        # Task 2: Data transformation
        transform = BashOperator(
            task_id='transform_data',
            bash_command='echo "Transforming data..."',
        )
        
        # Task 3: Load to warehorif
        load = BashOperator(
            task_id='load_data',
            bash_command='echo "Loading to data warehouse..."',
        )
        
        # Task 4: Run analytics
        analyze = PythonOperator(
            task_id='analyze',
            python_callable=detect_anomalies_batch,
        )
        
        # Task 5: Generatand report
        report = PythonOperator(
            task_id='report',
            python_callable=generate_daily_report,
        )
        
        # offinand ofpenofncies
        extract >> transform >> load >> analyze >> report
    
    logger.success("✅ Airflow DAGs defined")

else:
    logger.warning("Airflow not available - DAGs not created")

