requirements.txt
# LLM & AI Dependencies
openai==1.12.0
anthropic==0.18.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-anthropic==0.1.1
langchain-community==0.0.25

# Vector Stores & Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# Advanced ML
torch==2.2.0
transformers==4.37.2
scikit-learn==1.4.0
shap==0.44.0

# Existing dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
pyod==1.1.3
networkx==3.2.1
python-louvain==0.16
matplotlib==3.8.2
sqlalchemy==2.0.25
redis==5.0.1
kafka-python==2.0.2
httpx==0.26.0
requests==2.31.0
loguru==0.7.2
streamlit==1.30.0
plotly==5.18.0
faker==22.5.0
pytest==7.4.4

# Observability & Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Blockchain Analysis (optional)
web3==6.15.0
eth-utils==2.3.1

# ==================== DATA ENGINEERING ====================
kafka-python==2.0.2
confluent-kafka==2.3.0
apache-flink==1.18.0
pyspark==3.5.0
delta-spark==3.0.0
apache-airflow==2.7.3

# ==================== ADVANCED ML ====================
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
mlflow==2.8.1
optuna==3.4.0
hyperopt==0.2.7
torch-geometric==2.4.0

# ==================== LLM ADVANCED ====================
langgraph==0.0.26
dspy-ai==2.3.0
peft==0.7.0
bitsandbytes==0.41.3

# ==================== DATABASES ====================
neo4j==5.14.0
clickhouse-driver==0.2.6
psycopg2-binary==2.9.9

# ==================== ENTITY RESOLUTION ====================
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
dedupe==2.0.23
recordlinkage==0.15

# ==================== MODEL MONITORING ====================
evidently==0.4.11
whylabs==1.3.7
alibi-detect==0.11.4

# ==================== OBSERVABILITY ====================
datadog==0.48.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# ==================== ALERTING ====================
celery==5.3.4
flower==2.0.1

# ==================== UTILS ====================
tqdm==4.66.1
colorlog==6.8.0
